More work on the scanner
Patrick and Debs particularly, run this scanner and see if you can
figure out why it's not working exactly right.  I've got it to where it
finally loops through the whole file and doesn't get stuck in an
infinite loop.

However, it's getting a whitespace error (a check that I defined) on
every line at the end, and it shouldn't be.  I haven't figured out why.

It's also determining that a course name in line 7 is too long (but it's
not).  It is however, correctly finding a course too long in line 8.  So
take a look at this guys, if you don't mind.