Got it
I figured it out.  And as far as prereq025 goes, the scanner works.  Or
well, everything that isn't commented out works as it should.  Wasn't
even an end of line problem I was having.

Or well, it kind of was.  Either way, I solved it by checking for an end
of line (\r) marker at the end of each line.  But here it is.  I'll show
everyone at the meeting tomorrow.