EVERYONE READ DESCRIPTION AND FOLLOW INSTRUCTIONS
Alright so first of all, I integrated the scanner into the application.
The admin now can choose to schedule the course from the home page links
and they are prompted for how they want to prioritize faculty members,
then they schedule the classes.

ALSO EVERYONE DO THIS
I have cleaned up the git folder.
If you are running the project outside of the git folder, THEN YOU ARE
REQUIRED TO DO THIS.
Go through this list of deletes on your left and delete all these files
inside your localhost project folder. **Make sure and follow the each
path correctly**. I do not want to see these here again, and this is the
only way to prevent it from happening again for sure.

Thanks guys