Read duh-script-shun
Alright so.  All the "deleted" files is me simply renaming three
folders, so that's all good.

Next, you'll noticed the "Printable Listings" folder.  There's 4 files
in there listing a lot of things we need to print out in a (hopefully)
neat, organized fashion.  One of those files is "Still to do".  It lists
files that I was unclear on how to organize and/or what to associate
them with.  I need the rest of you to take care of those.

Finally, if you print anything from these files, find it in the
appropriate file and put a note next to it specifying that it has been
printed.

My next task is to compile a list of everything the scheduling algorithm
accounts for.  No clue if I'll get all of it so I'll probably need help
on it too.